requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
Django @ file:///tmp/build/80754af9/django_1606860386887/work
django-progressive-web-app==0.1.1
django-pwa==1.0.10
django-sslserver==0.22
djangorestframework==3.12.2
idna==2.10
Markdown==3.3.3
oauthlib==3.1.0
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
urllib3==1.26.2
wincertstore==0.2
